partial rejig
